Read SQL numeric data as string instead

+++ src/internal/pachsql/util.go
@@ -51,13 +51,12 @@ func SplitTableSchema(driver string, tablePath string) (schemaName string, table
 // TestRow is the type of a row in the test table
 // struct tag: sql:"<column_name>,<data_type>,<table_constraint>"
 type TestRow struct {
-	Id int `column:"c_id" dtype:"INT" constraint:"PRIMARY KEY NOT NULL"`
+	Id int16 `column:"c_id" dtype:"SMALLINT" constraint:"PRIMARY KEY NOT NULL"`
 
-	Smallint int16   `column:"c_smallint" dtype:"SMALLINT" constraint:"NOT NULL"`
-	Int      int32   `column:"c_int" dtype:"INT" constraint:"NOT NULL"`
-	Bigint   int64   `column:"c_bigint" dtype:"BIGINT" constraint:"NOT NULL"`
-	Float    float32 `column:"c_float" dtype:"FLOAT" constraint:"NOT NULL"`
-	// Numeric      float64   `column:"c_numeric" dtype:"NUMERIC" constraint:"NOT NULL"`
+	Smallint     int16     `column:"c_smallint" dtype:"SMALLINT" constraint:"NOT NULL"`
+	Int          int32     `column:"c_int" dtype:"INT" constraint:"NOT NULL"`
+	Bigint       int64     `column:"c_bigint" dtype:"BIGINT" constraint:"NOT NULL"`
+	Float        float32   `column:"c_float" dtype:"FLOAT" constraint:"NOT NULL"`
 	NumericInt   int64     `column:"c_numeric_int" dtype:"NUMERIC(20,0)" constraint:"NOT NULL"`
 	NumericFloat float64   `column:"c_numeric_float" dtype:"NUMERIC(20,2)" constraint:"NOT NULL"`
 	Varchar      string    `column:"c_varchar" dtype:"VARCHAR(100)" constraint:"NOT NULL"`
@@ -67,10 +66,8 @@ type TestRow struct {
 	IntNull      sql.NullInt32   `column:"c_int_null" dtype:"INT" constraint:"NULL"`
 	BigintNull   sql.NullInt64   `column:"c_bigint_null" dtype:"BIGINT" constraint:"NULL"`
 	FloatNull    sql.NullFloat64 `column:"c_float_null" dtype:"FLOAT" constraint:"NULL"`
-	// NumericIntNull   sql.NullInt64   `column:"c_numeric_int_null" dtype:"NUMERIC(20,0)" constraint:"NOT NULL"`
-	// NumericFloatNull sql.NullFloat64 `column:"c_decimal_null" dtype:"NUMERIC(20,2)" constraint:"NOT NULL"`
-	VarcharNull sql.NullString `column:"c_varchar_null" dtype:"VARCHAR(100)" constraint:"NULL"`
-	TimeNull    sql.NullTime   `column:"c_time_null" dtype:"TIMESTAMP" constraint:"NULL"`
+	VarcharNull  sql.NullString  `column:"c_varchar_null" dtype:"VARCHAR(100)" constraint:"NULL"`
+	TimeNull     sql.NullTime    `column:"c_time_null" dtype:"TIMESTAMP" constraint:"NULL"`
 }
 
 // CreateTestTable creates a test table at name in the database
@@ -97,7 +94,7 @@ func GenerateTestData(db *DB, tableName string, n int) error {
 	insertStatement := fmt.Sprintf("INSERT INTO test_data %s VALUES %s", formatColumns(row), formatValues(row, db))
 	for i := 0; i < n; i++ {
 		fz.Fuzz(&row)
-		row.Id = i
+		row.Id = int16(i)
 		row.Time = time.Now() // support mysql
 		if _, err := db.Exec(insertStatement, makeArgs(row)...); err != nil {
 			return errors.EnsureStack(err)
